requirements.txt
python-can>=4.3.0
python-uds>=1.0.0
cantools>=39.0.0
Flask>=3.0.0
flask-cors>=4.0.0
